DjangoTrain/pyproject.toml
[tool.poetry]
name = "djangotrain"
version = "0.1.0"
description = ""
authors = ["Tiago Duque <tiago.duque@beon.tech>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.3"
black = {extras = ["d"], version = "^23.7.0"}
djangorestframework = "^3.14.0"
django-cors-headers = "^4.2.0"
djangorestframework-simplejwt = "^5.2.2"
channels = "^4.0.0"
channels-redis = "^4.1.0"
djangochannelsrestframework = "^1.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
